Fix DMA readings, ADC2 failing after first DMA activation
The first activation of DMA works with ADC1 and ADC2
correctly, the second, ADC2 fails. To fix that, the
reset of the ADC wast't enough, so the reset of the
whole system is being made after each DMA activation.

A different solution may be reset the ADC and the DMA,
but it hasn't bein tested since the current setup is
working, and the tests with the ADC have taken signifi-
cant time.